#197: Move filters and clean up fetching logic
