+++ src/app/parcels/ParcelsPage.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-import Table, { Row, SortOptions, TableHeaders, SortState } from "@/components/Tables/Table";
-import React, { Suspense, useEffect, useRef, useState } from "react";
+import Table, { Row, SortOptions, SortState, TableHeaders } from "@/components/Tables/Table";
+import React, { Suspense, useCallback, useEffect, useRef, useState } from "react";
 import styled, { useTheme } from "styled-components";
 import { ParcelsTableRow } from "@/app/parcels/getParcelsTableData";
 import { formatDatetimeAsDate } from "@/app/parcels/getExpandedParcelDetails";
-import { DateRangeState } from "@/components/DateRangeInputs/DateRangeInputs";
 import FlaggedForAttentionIcon from "@/components/Icons/FlaggedForAttentionIcon";
 import PhoneIcon from "@/components/Icons/PhoneIcon";
 import CongestionChargeAppliesIcon from "@/components/Icons/CongestionChargeAppliesIcon";
@@ -21,31 +20,28 @@ import ActionAndStatusButtons from "@/app/parcels/ActionBar/ActionAndStatusButto
 import { ButtonsDiv, Centerer, ContentDiv, OutsideDiv } from "@/components/Modal/ModalFormStyles";
 import LinkButton from "@/components/Buttons/LinkButton";
 import supabase from "@/supabaseClient";
-import {
-    CollectionCentresOptions,
-    getParcelIds,
-    getParcelsByIds,
-    getParcelsDataAndCount,
-} from "./fetchParcelTableData";
+import { getParcelIds, getParcelsByIds, getParcelsDataAndCount } from "./fetchParcelTableData";
 import dayjs from "dayjs";
-import { checklistFilter } from "@/components/Tables/ChecklistFilter";
 import { Filter, PaginationType } from "@/components/Tables/Filters";
 import { saveParcelStatus } from "./ActionBar/Statuses";
 import { useRouter, useSearchParams } from "next/navigation";
-import { PostgrestFilterBuilder } from "@supabase/postgrest-js";
-import { Database } from "@/databaseTypesFile";
 import { buildTextFilter } from "@/components/Tables/TextFilter";
-import { dateFilter } from "@/components/Tables/DateFilter";
 import { CircularProgress, Paper } from "@mui/material";
 import { logErrorReturnLogId } from "@/logger/logger";
 import { DatabaseError } from "@/app/errorClasses";
 import { ErrorSecondaryText } from "../errorStylingandMessages";
 import { subscriptionStatusRequiresErrorMessage } from "@/common/subscriptionStatusRequiresErrorMessage";
-
-interface packingSlotOptionsSet {
-    key: string;
-    value: string;
-}
+import {
+    buildDateFilter,
+    buildDeliveryCollectionFilter,
+    buildLastStatusFilter,
+    buildPackingSlotFilter,
+    familySearch,
+    fullNameSearch,
+    phoneSearch,
+    postcodeSearch,
+    voucherSearch,
+} from "@/app/parcels/parcelsTableFilters";
 
 export const parcelTableHeaderKeysAndLabels: TableHeaders<ParcelsTableRow> = [
     ["iconsColumn", "Flags"],
@@ -217,208 +213,6 @@ function getSelectedParcelCountMessage(numberOfSelectedParcels: number): string 
 
 const parcelIdParam = "parcelId";
 
-const fullNameSearch = (
-    query: PostgrestFilterBuilder<Database["public"], any, any>,
-    state: string
-): PostgrestFilterBuilder<Database["public"], any, any> => {
-    return query.ilike("client_full_name", `%${state}%`);
-};
-
-const postcodeSearch = (
-    query: PostgrestFilterBuilder<Database["public"], any, any>,
-    state: string
-): PostgrestFilterBuilder<Database["public"], any, any> => {
-    return query.ilike("client_address_postcode", `%${state}%`);
-};
-
-const familySearch = (
-    query: PostgrestFilterBuilder<Database["public"], any, any>,
-    state: string
-): PostgrestFilterBuilder<Database["public"], any, any> => {
-    if (state === "") {
-        return query;
-    }
-    if ("single".includes(state.toLowerCase())) {
-        return query.lte("family_count", 1);
-    }
-    if ("family of".includes(state.toLowerCase())) {
-        return query.gte("family_count", 2);
-    }
-    const stateAsNumber = Number(state);
-    if (Number.isNaN(stateAsNumber) || stateAsNumber === 0) {
-        return query.eq("family_count", -1);
-    }
-    if (stateAsNumber >= 10) {
-        return query.gte("family_count", 10);
-    }
-    if (stateAsNumber === 1) {
-        return query.lte("family_count", 1);
-    }
-    return query.eq("family_count", Number(state));
-};
-
-const phoneSearch = (
-    query: PostgrestFilterBuilder<Database["public"], any, any>,
-    state: string
-): PostgrestFilterBuilder<Database["public"], any, any> => {
-    return query.ilike("client_phone_number", `%${state}%`);
-};
-
-const voucherSearch = (
-    query: PostgrestFilterBuilder<Database["public"], any, any>,
-    state: string
-): PostgrestFilterBuilder<Database["public"], any, any> => {
-    return query.ilike("voucher_number", `%${state}%`);
-};
-
-const buildDateFilter = (initialState: DateRangeState): Filter<ParcelsTableRow, DateRangeState> => {
-    const dateSearch = (
-        query: PostgrestFilterBuilder<Database["public"], any, any>,
-        state: DateRangeState
-    ): PostgrestFilterBuilder<Database["public"], any, any> => {
-        return query.gte("packing_date", state.from).lte("packing_date", state.to);
-    };
-    return dateFilter<ParcelsTableRow>({
-        key: "packingDate",
-        label: "",
-        methodConfig: { paginationType: PaginationType.Server, method: dateSearch },
-        initialState: initialState,
-    });
-};
-
-const buildDeliveryCollectionFilter = async (): Promise<Filter<ParcelsTableRow, string[]>> => {
-    const deliveryCollectionSearch = (
-        query: PostgrestFilterBuilder<Database["public"], any, any>,
-        state: string[]
-    ): PostgrestFilterBuilder<Database["public"], any, any> => {
-        return query.in("collection_centre_acronym", state);
-    };
-
-    const keySet = new Set();
-    const { data, error } = await supabase
-        .from("parcels_plus")
-        .select("collection_centre_name, collection_centre_acronym");
-    if (error) {
-        const logId = await logErrorReturnLogId(
-            "Error with fetch: Collection centre filter options",
-            error
-        );
-        throw new DatabaseError("fetch", "collection centre filter options", logId);
-    }
-    const optionsResponse = data ?? [];
-    const optionsSet: CollectionCentresOptions[] = optionsResponse.reduce<
-        CollectionCentresOptions[]
-    >((filteredOptions, row) => {
-        if (
-            row?.collection_centre_acronym &&
-            row.collection_centre_name &&
-            !keySet.has(row.collection_centre_acronym)
-        ) {
-            keySet.add(row.collection_centre_acronym);
-            filteredOptions.push({
-                name: row.collection_centre_name,
-                acronym: row.collection_centre_acronym,
-            });
-        }
-        return filteredOptions.sort();
-    }, []);
-
-    return checklistFilter<ParcelsTableRow>({
-        key: "deliveryCollection",
-        filterLabel: "Collection",
-        itemLabelsAndKeys: optionsSet.map((option) => [option!.name, option!.acronym]),
-        initialCheckedKeys: optionsSet.map((option) => option!.acronym),
-        methodConfig: { paginationType: PaginationType.Server, method: deliveryCollectionSearch },
-    });
-};
-
-const buildLastStatusFilter = async (): Promise<Filter<ParcelsTableRow, string[]>> => {
-    const lastStatusSearch = (
-        query: PostgrestFilterBuilder<Database["public"], any, any>,
-        state: string[]
-    ): PostgrestFilterBuilder<Database["public"], any, any> => {
-        if (state.includes("None")) {
-            return query.or(
-                `last_status_event_name.is.null,last_status_event_name.in.(${state.join(",")})`
-            );
-        } else {
-            return query.in("last_status_event_name", state);
-        }
-    };
-
-    const keySet = new Set();
-    const { data, error } = await supabase.from("status_order").select("event_name");
-    if (error) {
-        const logId = await logErrorReturnLogId("Error with fetch: Last status filter options");
-        throw new DatabaseError("fetch", "last status filter options", logId);
-    }
-    const optionsResponse = data ?? [];
-    const optionsSet: string[] = optionsResponse.reduce<string[]>((filteredOptions, row) => {
-        if (row.event_name && !keySet.has(row.event_name)) {
-            keySet.add(row.event_name);
-            filteredOptions.push(row.event_name);
-        }
-        return filteredOptions.sort();
-    }, []);
-    data && optionsSet.push("None");
-
-    return checklistFilter<ParcelsTableRow>({
-        key: "lastStatus",
-        filterLabel: "Last Status",
-        itemLabelsAndKeys: optionsSet.map((value) => [value, value]),
-        initialCheckedKeys: optionsSet.filter((option) => option !== "Request Deleted"),
-        methodConfig: { paginationType: PaginationType.Server, method: lastStatusSearch },
-    });
-};
-
-const buildPackingSlotFilter = async (): Promise<Filter<ParcelsTableRow, string[]>> => {
-    const packingSlotSearch = (
-        query: PostgrestFilterBuilder<Database["public"], any, any>,
-        state: string[]
-    ): PostgrestFilterBuilder<Database["public"], any, any> => {
-        return query.in("packing_slot_name", state);
-    };
-
-    const keySet = new Set();
-
-    const { data, error } = await supabase.from("packing_slots").select("name, is_shown");
-    if (error) {
-        const logId = await logErrorReturnLogId(
-            "Error with fetch: Packing slot filter options",
-            error
-        );
-        throw new DatabaseError("fetch", "packing slot filter options", logId);
-    }
-
-    const optionsResponse = data ?? [];
-
-    const optionsSet = optionsResponse.reduce<packingSlotOptionsSet[]>((filteredOptions, row) => {
-        if (!row.name || keySet.has(row.name)) {
-            return filteredOptions;
-        }
-
-        if (row.is_shown) {
-            keySet.add(row.name);
-            filteredOptions.push({ key: row.name, value: row.name });
-        } else {
-            keySet.add(row.name);
-            filteredOptions.push({ key: row.name, value: `${row.name} (inactive)` });
-        }
-
-        return filteredOptions;
-    }, []);
-
-    optionsSet.sort();
-
-    return checklistFilter<ParcelsTableRow>({
-        key: "packingSlot",
-        filterLabel: "Packing Slot",
-        itemLabelsAndKeys: optionsSet.map((option) => [option.value, option.key]),
-        initialCheckedKeys: optionsSet.map((option) => option.key),
-        methodConfig: { paginationType: PaginationType.Server, method: packingSlotSearch },
-    });
-};
-
 const ParcelsPage: React.FC<{}> = () => {
     const [isLoading, setIsLoading] = useState(true);
     const [parcelsDataPortion, setParcelsDataPortion] = useState<ParcelsTableRow[]>([]);
@@ -456,6 +250,22 @@ const ParcelsPage: React.FC<{}> = () => {
 
     const selectedParcelMessage = getSelectedParcelCountMessage(checkedParcelIds.length);
 
+    const fetchAndSetClientIdForSelectedParcel = async (): Promise<void> => {
+        const { data, error } = await supabase
+            .from("parcels")
+            .select("client_id")
+            .eq("primary_key", parcelId)
+            .single();
+
+        if (error) {
+            void logErrorReturnLogId("Error fetching clientId from database", { error });
+            return;
+        }
+
+        const fetchedClientId = data.client_id;
+        setClientIdForSelectedParcel(fetchedClientId);
+    };
+
     useEffect(() => {
         if (parcelId) {
             setSelectedParcelId(parcelId);
@@ -464,21 +274,6 @@ const ParcelsPage: React.FC<{}> = () => {
     }, [parcelId]);
 
     useEffect(() => {
-        const fetchAndSetClientIdForSelectedParcel = async (): Promise<void> => {
-            const { data, error } = await supabase
-                .from("parcels")
-                .select("client_id")
-                .eq("primary_key", parcelId)
-                .single();
-
-            if (error) {
-                void logErrorReturnLogId("Error fetching clientId from database", { error });
-                return;
-            }
-
-            const fetchedClientId = data.client_id;
-            setClientIdForSelectedParcel(fetchedClientId);
-        };
         setClientIdForSelectedParcel(null);
         void fetchAndSetClientIdForSelectedParcel();
     }, [parcelId]);
@@ -544,139 +339,98 @@ const ParcelsPage: React.FC<{}> = () => {
         })();
     }, []);
 
-    useEffect(() => {
-        if (!areFiltersLoadingForFirstTime) {
-            const allFilters = [...primaryFilters, ...additionalFilters];
+    const fetchAndDisplayParcelsData = useCallback(async (): Promise<void> => {
+        const allFilters = [...primaryFilters, ...additionalFilters];
+
+        if (parcelsTableFetchAbortController.current) {
+            parcelsTableFetchAbortController.current.abort("stale request");
+        }
+
+        parcelsTableFetchAbortController.current = new AbortController();
+
+        if (parcelsTableFetchAbortController.current) {
+            setErrorMessage(null);
             setIsLoading(true);
-            if (parcelsTableFetchAbortController.current) {
-                parcelsTableFetchAbortController.current.abort("stale request");
-            }
 
-            parcelsTableFetchAbortController.current = new AbortController();
-
-            if (parcelsTableFetchAbortController.current) {
-                setErrorMessage(null);
-                getParcelsDataAndCount(
+            try {
+                const { data, count } = await getParcelsDataAndCount(
                     supabase,
                     allFilters,
                     sortState,
                     parcelsTableFetchAbortController.current.signal,
                     startPoint,
                     endPoint
-                )
-                    .then(({ data, count }) => {
-                        setParcelsDataPortion(data);
-                        setFilteredParcelCount(count);
-                    })
-                    .catch((error) => {
-                        if (error instanceof DatabaseError) {
-                            setErrorMessage(error.message);
-                        }
-                    })
-                    .finally(() => {
-                        parcelsTableFetchAbortController.current = null;
-                        setIsLoading(false);
-                    });
+                );
+                setParcelsDataPortion(data);
+                setFilteredParcelCount(count);
+            } catch (error) {
+                if (error instanceof DatabaseError) {
+                    setErrorMessage(error.message);
+                }
+            } finally {
+                parcelsTableFetchAbortController.current = null;
+                setIsLoading(false);
             }
         }
-    }, [
-        startPoint,
-        endPoint,
-        primaryFilters,
-        additionalFilters,
-        sortState,
-        areFiltersLoadingForFirstTime,
-    ]);
+    }, [additionalFilters, endPoint, primaryFilters, sortState, startPoint]);
 
     useEffect(() => {
-        // This requires that the DB parcels, events, families, clients and collection_centres tables have Realtime turned on
-        if (!areFiltersLoadingForFirstTime) {
-            const allFilters = [...primaryFilters, ...additionalFilters];
-            const loadCountAndDataWithTimer = (): void => {
-                if (fetchParcelsTimer.current) {
-                    clearTimeout(fetchParcelsTimer.current);
-                    fetchParcelsTimer.current = null;
-                }
-
-                setIsLoading(true);
-                fetchParcelsTimer.current = setTimeout(() => {
-                    if (parcelsTableFetchAbortController.current) {
-                        parcelsTableFetchAbortController.current.abort("stale request");
-                    }
-
-                    parcelsTableFetchAbortController.current = new AbortController();
-
-                    if (parcelsTableFetchAbortController.current) {
-                        setErrorMessage(null);
-                        getParcelsDataAndCount(
-                            supabase,
-                            allFilters,
-                            sortState,
-                            parcelsTableFetchAbortController.current.signal,
-                            startPoint,
-                            endPoint
-                        )
-                            .then(({ data, count }) => {
-                                setParcelsDataPortion(data);
-                                setFilteredParcelCount(count);
-                            })
-                            .catch((error) => {
-                                if (error instanceof DatabaseError) {
-                                    setErrorMessage(error.message);
-                                }
-                            })
-                            .finally(() => {
-                                parcelsTableFetchAbortController.current = null;
-                                setIsLoading(false);
-                            });
-                    }
-                }, 500);
-            };
-
-            const subscriptionChannel = supabase
-                .channel("parcels-table-changes")
-                .on(
-                    "postgres_changes",
-                    { event: "*", schema: "public", table: "parcels" },
-                    loadCountAndDataWithTimer
-                )
-                .on(
-                    "postgres_changes",
-                    { event: "*", schema: "public", table: "events" },
-                    loadCountAndDataWithTimer
-                )
-                .on(
-                    "postgres_changes",
-                    { event: "*", schema: "public", table: "families" },
-                    loadCountAndDataWithTimer
-                )
-                .on(
-                    "postgres_changes",
-                    { event: "*", schema: "public", table: "collection_centres" },
-                    loadCountAndDataWithTimer
-                )
-                .on(
-                    "postgres_changes",
-                    { event: "*", schema: "public", table: "clients" },
-                    loadCountAndDataWithTimer
-                )
-                .subscribe((status, err) => {
-                    subscriptionStatusRequiresErrorMessage(status, err, "website_data") &&
-                        setErrorMessage("Error fetching data, please reload");
-                });
-
-            return () => {
-                supabase.removeChannel(subscriptionChannel);
-            };
+        if (areFiltersLoadingForFirstTime) {
+            return;
         }
-    }, [
-        endPoint,
-        startPoint,
-        primaryFilters,
-        additionalFilters,
-        sortState,
-        areFiltersLoadingForFirstTime,
-    ]);
+
+        void fetchAndDisplayParcelsData();
+    }, [areFiltersLoadingForFirstTime, fetchAndDisplayParcelsData]);
+
+    const loadCountAndDataWithTimer = (): void => {
+        if (fetchParcelsTimer.current) {
+            clearTimeout(fetchParcelsTimer.current);
+            fetchParcelsTimer.current = null;
+        }
+
+        setIsLoading(true);
+        fetchParcelsTimer.current = setTimeout(() => {
+            void fetchAndDisplayParcelsData();
+        }, 500);
+    };
+
+    useEffect(() => {
+        const subscriptionChannel = supabase
+            .channel("parcels-table-changes")
+            .on(
+                "postgres_changes",
+                { event: "*", schema: "public", table: "parcels" },
+                loadCountAndDataWithTimer
+            )
+            .on(
+                "postgres_changes",
+                { event: "*", schema: "public", table: "events" },
+                loadCountAndDataWithTimer
+            )
+            .on(
+                "postgres_changes",
+                { event: "*", schema: "public", table: "families" },
+                loadCountAndDataWithTimer
+            )
+            .on(
+                "postgres_changes",
+                { event: "*", schema: "public", table: "collection_centres" },
+                loadCountAndDataWithTimer
+            )
+            .on(
+                "postgres_changes",
+                { event: "*", schema: "public", table: "clients" },
+                loadCountAndDataWithTimer
+            )
+            .subscribe((status, err) => {
+                subscriptionStatusRequiresErrorMessage(status, err, "website_data") &&
+                    setErrorMessage("Error fetching data, please reload");
+            });
+
+        return () => {
+            void supabase.removeChannel(subscriptionChannel);
+        };
+    });
 
     const selectOrDeselectRow = (parcelId: string): void => {
         setCheckedParcelIds((currentIndices) => {
